+++ Day_03/Part_02.cs
@@ -3,7 +3,7 @@ using HelperLibrary;
 
 namespace Day_03 {
     public class Part_02 : IDay {
-        public int Solve(string[] pInput) {
+        public long Solve(string[] pInput) {
             var pattern = @"(mul\(\d{1,3},\d{1,3}\))|(don't|do)";
 
             var matches = pInput.SelectMany(line => Regex.Matches(line, pattern, RegexOptions.Multiline).Select(match => match.Value ));
